+++ packages/adminui-vue/CHANGELOG.md
@@ -6,8 +6,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+ - `UMasterDetailView`: form preview mode. Also added:
+   - `contextMenuPreviewButton` slot
+   - `dropdownMenuPreviewButton` slot
+
  - `UToolbar`: 'toolbarInfoRow' slot for additional information about record in toolbar
- under the dates of creation and modification
+   under the dates of creation and modification
+
 ### Changed
 
 ### Deprecated